Beatport Discography Shortcode
Version 1.0

Shortcode to display artist discography.

The discography is a feed using Beatport API and rendered in one or two
columns depending on the width of the page.

The way to use this plugin is simply add a new page or post and on the
new page you will see a button with Beatport logo on the toolbar of the
editor.

If you click the logo, it opens a modal page with a form to enter
parameters.

On this first version the only parameter being used is artist name
which has to be exactly as it appears on beatport.